+++ src/utils/array/matrix/RotateLeft.js
@@ -18,7 +18,7 @@ var RotateMatrix = require('./RotateMatrix');
  */
 var RotateLeft = function (matrix)
 {
-    return RotateMatrix(matrix, -90);
+    return RotateMatrix(matrix, 90);
 };
 
 module.exports = RotateLeft;