+++ src/utils/array/matrix/RotateRight.js
@@ -18,7 +18,7 @@ var RotateMatrix = require('./RotateMatrix');
  */
 var RotateRight = function (matrix)
 {
-    return RotateMatrix(matrix, 90);
+    return RotateMatrix(matrix, -90);
 };
 
 module.exports = RotateRight;